+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Clever Ads Solutions Unity Plugin Change Log
 
+## [3.1.8] - 2023-05-31
+### Dependencies
+- [Android] Wraps [3.1.8 SDK](https://github.com/cleveradssolutions/CAS-Android/releases)
+- [iOS] Wraps [3.1.8 SDK](https://github.com/cleveradssolutions/CAS-iOS/releases)
+
+## [3.1.7] - 2023-05-24
+### Dependencies
+- [Android] Wraps [3.1.7 SDK](https://github.com/cleveradssolutions/CAS-Android/releases)
+- [iOS] Wraps [3.1.7 SDK](https://github.com/cleveradssolutions/CAS-iOS/releases)
+
 ## [3.1.6] - 2023-05-17
 ### Dependencies
 - [Android] Wraps [3.1.6 SDK](https://github.com/cleveradssolutions/CAS-Android/releases)